+++ depthai_nodes/__init__.py
@@ -1,5 +1,4 @@
 from .message import *
-from .node import *
 from .utils import setup_logging
 
 __version__ = "0.1.2"
