+++ depthai_nodes/node/parsers/utils/keypoints.py
@@ -1,8 +1,4 @@
-from typing import List, Optional
-
 import numpy as np
-
-from depthai_nodes import Keypoint
 
 
 def normalize_keypoints(keypoints: np.ndarray, height: int, width: int) -> np.ndarray:
@@ -41,57 +37,3 @@ def normalize_keypoints(keypoints: np.ndarray, height: int, width: int) -> np.nd
     keypoints[:, 1] = keypoints[:, 1] / height
 
     return keypoints
-
-
-def transform_to_keypoints(
-    keypoints: np.ndarray, confidences: Optional[np.ndarray] = None
-) -> List[Keypoint]:
-    """Transforms an array representing keypoints into a list of Keypoint objects.
-
-    @param keypoints: Detected 2D or 3D keypoints of shape (N,2 or 3) meaning [...,[x, y],...] or [...,[x, y, z],...].
-    @type keypoints: np.ndarray
-    @param confidences: Confidence scores of the detected keypoints.
-    @type confidences: Optional[np.ndarray]
-
-    @return: List of Keypoint objects.
-    @rtype: List[Keypoint]
-    """
-    if not isinstance(keypoints, np.ndarray):
-        raise ValueError("Keypoints must be a numpy array.")
-    if len(keypoints.shape) != 2:
-        raise ValueError(
-            f"Keypoints must be of shape (N, 2). Got shape {keypoints.shape}."
-        )
-    if keypoints.shape[1] != 2 and keypoints.shape[1] != 3:
-        raise ValueError(
-            f"Keypoints must be of shape (N, 2) or (N, 3). Got shape {keypoints.shape}."
-        )
-
-    if confidences is not None:
-        if not isinstance(confidences, np.ndarray):
-            raise ValueError("Confidences must be a numpy array.")
-        if len(confidences.shape) != 1:
-            raise ValueError(
-                f"Confidences must be of shape (N,). Got shape {confidences.shape}."
-            )
-        if len(confidences) != len(keypoints):
-            raise ValueError(
-                f"Confidences should have same length as keypoints, got {len(confidences)} confidences and {len(keypoints)} keypoints."
-            )
-
-    dim = keypoints.shape[1]
-
-    keypoints_list = []
-    for i, keypoint in enumerate(keypoints):
-        kp = Keypoint()
-        kp.x = keypoint[0]
-        kp.y = keypoint[1]
-
-        if dim == 3:
-            kp.z = keypoint[2]
-        if confidences is not None:
-            kp.confidence = confidences[i]
-
-        keypoints_list.append(kp)
-
-    return keypoints_list
