+++ depthai_nodes/node/tiles_patcher.py
@@ -2,8 +2,9 @@ from typing import List
 
 import depthai as dai
 
-from depthai_nodes.node import Tiling
 from depthai_nodes.node.utils import nms_detections
+
+from .tiling import Tiling
 
 
 class TilesPatcher(dai.node.HostNode):
